API

API

OnlineStats.ADADELTA

OnlineStats.ADAGRAD

OnlineStats.ADAM

OnlineStats.ADAMAX

OnlineStats.Bootstrap

OnlineStats.CovMatrix

OnlineStats.Diff

OnlineStats.Extrema

OnlineStats.FitBeta

OnlineStats.FitCategorical

OnlineStats.FitCauchy

OnlineStats.FitGamma

OnlineStats.FitLogNormal

OnlineStats.FitMultinomial

OnlineStats.FitMvNormal

OnlineStats.FitNormal

OnlineStats.HyperLogLog

OnlineStats.KMeans

OnlineStats.LinReg

OnlineStats.MSPIC

OnlineStats.MSPIF

OnlineStats.MV

OnlineStats.Mean

OnlineStats.Moments

OnlineStats.NADAM

OnlineStats.NSGD

OnlineStats.OHistogram

OnlineStats.OMASQ

OnlineStats.OMASQF

OnlineStats.OrderStats

OnlineStats.Quantiles

OnlineStats.ReservoirSample

OnlineStats.SGD

OnlineStats.SPI

OnlineStats.Series

OnlineStats.StatLearn

OnlineStats.StochasticLoss

OnlineStats.Sum

OnlineStats.Variance

OnlineStatsBase.AbstractSeries

OnlineStatsBase.EqualWeight

OnlineStatsBase.ExponentialWeight

OnlineStatsBase.HarmonicWeight

OnlineStatsBase.LearningRate

OnlineStatsBase.LearningRate2

OnlineStatsBase.McclainWeight

LearnBase.value

OnlineStats.maprows

OnlineStats.replicates

OnlineStats.stats

StatsBase.confint

StatsBase.fit!

OnlineStats.Bootstrap — Type.

Bootstrap(s::Series, nreps, d, f = value)

Online Statistical Bootstrapping.

Create nreps replicates of the OnlineStat in Series s. When fit! is called, each of the replicates will be updated rand(d) times. Standard choices for d are Distributions.Poisson(), [0, 2], etc. value(b) returns f mapped to the replicates.

Example

b = Bootstrap(Series(Mean()), 100, [0, 2])
fit!(b, randn(1000))
value(b)        # `f` mapped to replicates
mean(value(b))  # mean

source

OnlineStats.Diff — Type.

Diff()

Track the difference and the last value.

Example

s = Series(randn(1000), Diff())
value(s)

source

OnlineStats.FitCategorical — Type.

FitCategorical(T)

Fit a categorical distribution where the inputs are of type T.

Example

using Distributions
s = Series(rand(1:10, 1000), FitCategorical(Int))
value(s)

vals = ["small", "medium", "large"]
s = Series(rand(vals, 1000), FitCategorical(String))
value(s)

source

OnlineStats.MV — Type.

MV(p, o)

Track p univariate OnlineStats o

Example

y = randn(1000, 5)
o = MV(5, Mean())
s = Series(y, o)

source

OnlineStats.OHistogram — Type.

OHistogram(range)

Make a histogram with bins given by range. Uses left-closed bins.

Example

y = randn(100)
s = Series(y, OHistogram(-4:.1:4))
value(s)

source

OnlineStats.Quantiles — Type.

Quantiles(q = [.25, .5, .75])  # default algorithm is :MSPI
Quantiles{:SGD}(q = [.25, .5, .75])
Quantiles{:MSPI}(q = [.25, .5, .75])

Approximate quantiles via the specified algorithm (:SGD or :MSPI).

Example

s = Series(randn(10_000), Quantiles(.1:.1:.9)

source

OnlineStats.ReservoirSample — Type.

ReservoirSample(k)
ReservoirSample(k, Float64)

Reservoir sample of k items.

Example

o = ReservoirSample(k, Int)
s = Series(o)
fit!(s, 1:10000)

source

OnlineStats.Series — Type.

Series(stats...)
Series(data, stats...)
Series(weight, stats...)
Series(weight, data, stats...)

A Series is a container for a Weight and any number of OnlineStats. Updating the Series with fit!(s, data) will update the OnlineStats it holds according to its Weight.

Examples

Series(randn(100), Mean(), Variance())
Series(ExponentialWeight(.1), Mean())

s = Series(Mean())
fit!(s, randn(100))
s2 = Series(randn(123), Mean())
merge(s, s2)

source

OnlineStats.StatLearn — Type.

StatLearn(p, loss, penalty, λ, updater)

Fit a statistical learning model of p independent variables for a given loss, penalty, and λ. Arguments are:

loss: any Loss from LossFunctions.jl

penalty: any Penalty from PenaltyFunctions.jl.

λ: a Vector of element-wise regularization parameters

updater: SPGD(), ADAGRAD(), ADAM(), or ADAMAX()

Example

using LossFunctions, PenaltyFunctions
x = randn(100_000, 10)
y = x * linspace(-1, 1, 10) + randn(100_000)
o = StatLearn(10, L2DistLoss(), L1Penalty(), fill(.1, 10), SPGD())
s = Series(o)
fit!(s, x, y)
coef(o)
predict(o, x)

source

OnlineStats.StochasticLoss — Type.

    s = Series(randn(1000), StochasticLoss(QuantileLoss(.7)))

Minimize a loss (from LossFunctions.jl) using stochastic gradient descent.

Example

o1 = StochasticLoss(QuantileLoss(.7))  # approx. .7 quantile
o2 = StochasticLoss(L2DistLoss())      # approx. mean
o3 = StochasticLoss(L1DistLoss())      # approx. median
s = Series(randn(10_000), o1, o2, o3)

source

OnlineStats.Sum — Type.

Sum()

Track the overall sum.

Example

s = Series(randn(1000), Sum())
value(s)

source

OnlineStats.ADADELTA — Type.

ADADELTA(η = 1.0, ρ = .95)

ADADELTA ignores weight.

source

OnlineStats.ADAGRAD — Type.

ADAGRAD(η)

Adaptive (element-wise learning rate) SPGD with step size η

source

OnlineStats.ADAM — Type.

ADAM(η, α1, α2)

Adaptive Moment Estimation with step size η and momentum parameters α1, α2

source

OnlineStats.ADAMAX — Type.

ADAMAX(η, β1, β2)

ADAMAX with step size η and momentum parameters β1, β2

source

OnlineStats.CovMatrix — Type.

CovMatrix(d)

Covariance Matrix of d variables.

Example

y = randn(100, 5)
Series(y, CovMatrix(5))

source

OnlineStats.Extrema — Type.

Extrema()

Maximum and minimum.

Example

s = Series(randn(100), Extrema())
value(s)

source

OnlineStats.FitBeta — Type.

FitBeta()

Online parameter estimate of a Beta distribution (Method of Moments)

Example

using Distributions, OnlineStats
y = rand(Beta(3, 5), 1000)
s = Series(y, FitBeta())
Beta(value(s)...)

source

OnlineStats.FitCauchy — Type.

FitCauchy()

Online parameter estimate of a Cauchy distribution

Example

using Distributions
y = rand(Cauchy(0, 10), 10_000)
s = Series(y, FitCauchy())
Cauchy(value(s)...)

source

OnlineStats.FitGamma — Type.

FitGamma()

Online parameter estimate of a Gamma distribution (Method of Moments)

Example

using Distributions
y = rand(Gamma(5, 1), 1000)
s = Series(y, FitGamma())
Gamma(value(s)...)

source

OnlineStats.FitLogNormal — Type.

FitLogNormal()

Online parameter estimate of a LogNormal distribution (MLE)

Example

using Distributions
y = rand(LogNormal(3, 4), 1000)
s = Series(y, FitLogNormal())
LogNormal(value(s)...)

source

OnlineStats.FitMultinomial — Type.

FitMultinomial(p)

Online parameter estimate of a Multinomial distribution.

Example

using Distributions
y = rand(Multinomial(10, [.2, .2, .6]), 1000)
s = Series(y', FitMultinomial())
Multinomial(value(s)...)

source

OnlineStats.FitMvNormal — Type.

FitMvNormal(d)

Online parameter estimate of a d-dimensional MvNormal distribution (MLE)

Example

using Distributions
y = rand(MvNormal(zeros(3), eye(3)), 1000)
s = Series(y', FitMvNormal(3))

source

OnlineStats.FitNormal — Type.

FitNormal()

Online parameter estimate of a Normal distribution (MLE)

Example

using Distributions
y = rand(Normal(-3, 4), 1000)
s = Series(y, FitNormal())

source

OnlineStats.HyperLogLog — Type.

HyperLogLog(b)  # 4 ≤ b ≤ 16

Approximate count of distinct elements.

Example

s = Series(rand(1:10, 1000), HyperLogLog(12))

source

OnlineStats.KMeans — Type.

KMeans(p, k)

Approximate K-Means clustering of k clusters of p variables

Example

using OnlineStats, Distributions
d = MixtureModel([Normal(0), Normal(5)])
y = rand(d, 100_000, 1)
s = Series(y, LearningRate(.6), KMeans(1, 2))

source

OnlineStats.LinReg — Type.

LinReg(p)
LinReg(p, λ)

Create a linear regression object with p predictors and optional ridge (L2-regularization) parameter λ.

Example

x = randn(1000, 5)
y = x * linspace(-1, 1, 5) + randn(1000)
o = LinReg(5)
s = Series(o)
fit!(s, x, y)
coef(o)
predict(o, x)
coeftable(o)
vcov(o)
confint(o)

source

OnlineStats.MSPIC — Type.

Experimental: MSPI-constant

source

OnlineStats.MSPIF — Type.

Experimental: MSPI-full matrix

source

OnlineStats.Mean — Type.

Mean()

Univariate mean.

Example

s = Series(randn(100), Mean())
value(s)

source

OnlineStats.Moments — Type.

Moments()

First four non-central moments.

Example

s = Series(randn(1000), Moments(10))
value(s)

source

OnlineStats.NADAM — Type.

NADAM(η, α1, α2)

Adaptive Moment Estimation with step size η and momentum parameters α1, α2

source

OnlineStats.NSGD — Type.

NSGD(η, α)

Nesterov accelerated Stochastic Proximal Gradient Descent.

source

OnlineStats.OMASQ — Type.

Experimental: OMM-constant

source

OnlineStats.OMASQF — Type.

Experimental: OMM-full matrix

source

OnlineStats.OrderStats — Type.

OrderStats(b)

Average order statistics with batches of size b.

Example

s = Series(randn(1000), OrderStats(10))
value(s)

source

OnlineStats.SGD — Type.

SGD(η, α=0.0)

Stochastic Proximal Gradient Descent with step size η and momentum term α.

source

OnlineStats.SPI — Type.

Stochastic Proximal Iteration

source

OnlineStats.Variance — Type.

Variance()

Univariate variance.

Example

s = Series(randn(100), Variance())
value(s)

source

LearnBase.value — Method.

Map value to the stats field of a Series.

source

OnlineStats.maprows — Method.

maprows(f::Function, b::Integer, data...)

Map rows of data in batches of size b. Most usage is done through do blocks.

Example

s = Series(Mean())
maprows(10, randn(100)) do yi
    fit!(s, yi)
    info("nobs: $(nobs(s))")
end

source

OnlineStats.replicates — Method.

replicates(b)

Return the vector of replicates from Bootstrap b

source

OnlineStats.stats — Method.

Return the stats field of a Series.

source

StatsBase.confint — Function.

confint(b, coverageprob = .95)

Return a confidence interval for a Bootstrap b.

source

StatsBase.fit! — Function.

fit!(s, y)
fit!(s, y, w)

Update a Series s with more data y and optional weighting w.

Examples

y = randn(100)
w = rand(100)

s = Series(Mean())
fit!(s, y[1])        # one observation: use Series weight
fit!(s, y[1], w[1])  # one observation: override weight
fit!(s, y)           # multiple observations: use Series weight
fit!(s, y, w[1])     # multiple observations: override each weight with w[1]
fit!(s, y, w)        # multiple observations: y[i] uses weight w[i]

source

OnlineStatsBase.AbstractSeries — Type.

A container for a Weight and at least one OnlineStat

source

OnlineStatsBase.EqualWeight — Type.

EqualWeight()

Equally weighted observations

Weight at observation t is γ = 1 / t

source

OnlineStatsBase.ExponentialWeight — Type.

ExponentialWeight(λ::Real = 0.1)
ExponentialWeight(lookback::Integer)

Exponentially weighted observations (constant)

Weight at observation t is γ = λ

source

OnlineStatsBase.HarmonicWeight — Type.

HarmonicWeight(a = 10.0)

Decreases at a slow rate

Weight at observation t is γ = a / (a + t - 1)

source

OnlineStatsBase.LearningRate — Type.

LearningRate(r = .6)

Mainly for stochastic approximation types (QuantileSGD, QuantileMM etc.)

Decreases at a "slow" rate

Weight at observation t is γ = 1 / t ^ r

source

OnlineStatsBase.LearningRate2 — Type.

LearningRate2(c = .5)

Mainly for stochastic approximation types (QuantileSGD, QuantileMM etc.)

Decreases at a "slow" rate

Weight at observation t is γ = inv(1 + c * (t - 1))

source

OnlineStatsBase.McclainWeight — Type.

McclainWeight(ᾱ = 0.1)

"smoothed" version of BoundedEqualWeight

weights asymptotically approach ᾱ

Weight at observation t is γ(t-1) / (1 + γ(t-1) - ᾱ)

source