API

API

OnlineStats.ADADELTA

OnlineStats.ADAGRAD

OnlineStats.ADAM

OnlineStats.ADAMAX

OnlineStats.AdaptiveBins

OnlineStats.AugmentedSeries

OnlineStats.AutoCov

OnlineStats.BiasVec

OnlineStats.Bootstrap

OnlineStats.CStat

OnlineStats.CallFun

OnlineStats.Count

OnlineStats.CountMap

OnlineStats.CovMatrix

OnlineStats.Diff

OnlineStats.Extrema

OnlineStats.FitBeta

OnlineStats.FitCauchy

OnlineStats.FitGamma

OnlineStats.FitLogNormal

OnlineStats.FitMultinomial

OnlineStats.FitMvNormal

OnlineStats.FitNormal

OnlineStats.Group

OnlineStats.Hist

OnlineStats.HyperLogLog

OnlineStats.KMeans

OnlineStats.KnownBins

OnlineStats.Lag

OnlineStats.LinReg

OnlineStats.LinRegBuilder

OnlineStats.MSPI

OnlineStats.MSPI2

OnlineStats.MV

OnlineStats.Mean

OnlineStats.Moments

OnlineStats.NADAM

OnlineStats.NBClassifier

OnlineStats.NSGD

OnlineStats.OMAP

OnlineStats.OMAP2

OnlineStats.OMAS

OnlineStats.OMAS2

OnlineStats.OrderStats

OnlineStats.PQuantile

OnlineStats.Partition

OnlineStats.Quantile

OnlineStats.RMSPROP

OnlineStats.ReservoirSample

OnlineStats.SGD

OnlineStats.Series

OnlineStats.StatLearn

OnlineStats.Sum

OnlineStats.Variance

OnlineStatsBase.Bounded

OnlineStatsBase.EqualWeight

OnlineStatsBase.ExactStat

OnlineStatsBase.ExponentialWeight

OnlineStatsBase.HarmonicWeight

OnlineStatsBase.LearningRate

OnlineStatsBase.LearningRate2

OnlineStatsBase.McclainWeight

OnlineStatsBase.Scaled

OnlineStatsBase.StochasticStat

OnlineStatsBase.Weight

Base.merge

Base.merge!

OnlineStats.always

OnlineStats.mapblocks

OnlineStatsBase._fit!

StatsBase.confint

StatsBase.fit!

OnlineStats.ADADELTA — Type.

ADADELTA(ρ = .95)

ADADELTA ignores weight.

source

OnlineStats.ADAGRAD — Type.

ADAGRAD()

Adaptive (element-wise learning rate) stochastic gradient descent.

source

OnlineStats.ADAM — Type.

ADAM(α1 = .99, α2 = .999)

Adaptive Moment Estimation with momentum parameters α1 and α2.

source

OnlineStats.ADAMAX — Type.

ADAMAX(η, β1 = .9, β2 = .999)

ADAMAX with step size η and momentum parameters β1, β2

source

OnlineStats.AdaptiveBins — Type.

Calculate a histogram adaptively.

Ref: http://www.jmlr.org/papers/volume11/ben-haim10a/ben-haim10a.pdf

source

OnlineStats.AugmentedSeries — Type.

AugmentedSeries(s::Series; filter = x->true, transform = identity)

Wrapper around a Series so that for new data, fitting occurs on transform(data), but only if filter(data) == true.

source

OnlineStats.AutoCov — Type.

AutoCov(b, T = Float64)

Calculate the auto-covariance/correlation for lags 0 to b for a data stream of type T.

Example

y = cumsum(randn(100))
o = AutoCov(5)
Series(y, o)
autocov(o)
autocor(o)

source

OnlineStats.Bootstrap — Type.

Bootstrap(o::OnlineStat, nreps = 100, d = [0, 2])

Online statistical bootstrap. Create nreps replicates of o. For each call to fit!, a replicate will be updated rand(d) times.

Example

o = Bootstrap(Variance())
Series(randn(1000), o)
confint(o)

source

OnlineStats.CStat — Type.

CStat(stat)

Track a univariate OnlineStat for complex numbers. A copy of stat is made to separately track the real and imaginary parts.

Example

y = randn(100) + randn(100)im
Series(y, CStat(Mean()))

source

OnlineStats.CallFun — Type.

CallFun(o::OnlineStat, f::Function)

Call f(o) every time the OnlineStat o gets updated.

Example

Series(randn(5), CallFun(Mean(), info))

source

OnlineStats.Count — Type.

Count()

The number of things observed.

source

OnlineStats.CountMap — Type.

CountMap(T)

Maintain a dictionary mapping unique values to its number of occurrences. Ignores weight and is equivalent to StatsBase.countmap.

Example

s = Series(rand(1:10, 1000), CountMap(Int))
value(s)[1]

vals = ["small", "medium", "large"]
o = CountMap(String)
s = Series(rand(vals, 1000), o)
value(o)

source

OnlineStats.CovMatrix — Type.

CovMatrix(d)

Covariance Matrix of d variables. Principal component analysis can be performed using eigen decomposition of the covariance or correlation matrix.

Example

y = randn(100, 5)
o = CovMatrix(5)
Series(y, o)

# PCA
evals, evecs = eig(cor(o))

source

OnlineStats.Diff — Type.

Diff()

Track the difference and the last value.

Example

s = Series(randn(1000), Diff())
value(s)

source

OnlineStats.Extrema — Type.

Extrema()

Maximum and minimum.

Example

s = Series(randn(100), Extrema())
value(s)

source

OnlineStats.FitBeta — Type.

FitBeta()

Online parameter estimate of a Beta distribution (Method of Moments).

using Distributions, OnlineStats
y = rand(Beta(3, 5), 1000)
o = FitBeta()
s = Series(y, o)
Beta(value(o)...)

source

OnlineStats.FitCauchy — Type.

FitCauchy(alg = SGD())

Approximate parameter estimation of a Cauchy distribution. Estimates are based on quantiles, so that alg will be passed to Quantile.

using Distributions
y = rand(Cauchy(0, 10), 10_000)
o = FitCauchy(SGD())
s = Series(y, o)
Cauchy(value(o)...)

source

OnlineStats.FitGamma — Type.

FitGamma()

Online parameter estimate of a Gamma distribution (Method of Moments).

using Distributions
y = rand(Gamma(5, 1), 1000)
o = FitGamma()
s = Series(y, o)
Gamma(value(o)...)

source

OnlineStats.FitLogNormal — Type.

FitLogNormal()

Online parameter estimate of a LogNormal distribution (MLE).

using Distributions
y = rand(LogNormal(3, 4), 1000)
o = FitLogNormal()
s = Series(y, o)
LogNormal(value(o)...)

source

OnlineStats.FitMultinomial — Type.

FitMultinomial(p)

Online parameter estimate of a Multinomial distribution. The sum of counts does not need to be consistent across observations. Therefore, the n parameter of the Multinomial distribution is returned as 1.

using Distributions
y = rand(Multinomial(10, [.2, .2, .6]), 1000)
o = FitMultinomial(3)
s = Series(y', o)
Multinomial(value(o)...)

source

OnlineStats.FitMvNormal — Type.

FitMvNormal(d)

Online parameter estimate of a d-dimensional MvNormal distribution (MLE).

using Distributions
y = rand(MvNormal(zeros(3), eye(3)), 1000)
o = FitMvNormal(3)
s = Series(y', o)
MvNormal(value(o)...)

source

OnlineStats.FitNormal — Type.

FitNormal()

Online parameter estimate of a Normal distribution (MLE).

using Distributions
y = rand(Normal(-3, 4), 1000)
o = FitNormal()
s = Series(y, o)
Normal(value(o)...)

source

OnlineStats.Group — Type.

Group(stats...)

Create an ExactStat{1} from several OnlineStat{0}s. For a new observation y, y[i] is sent to stats[i]. This is designed for working with data of different variable types.

Example

y = [randn(100) rand(["a", "b"], 100)]

o = Group(Mean(), CountMap(String))

Series(y, o)

value(o)

source

OnlineStats.Hist — Type.

Hist(r::Range)
Hist(b::Int)

Calculate a histogram over bin edges fixed as r or adaptively find the best b bins. The two options use KnownBins and AdaptiveBins, respectively. KnownBins is much faster, but requires the range of the data to be known before it is observed. Hist objects can be used to return approximate summary statistics of the data.

Example

o = Hist(-5:.1:5)
y = randn(1000)
Series(y, o)

# approximate summary statistics
mean(o)
var(o)
std(o)
median(o)
extrema(o)
quantile(o)

source

OnlineStats.HyperLogLog — Type.

HyperLogLog(b)  # 4 ≤ b ≤ 16

Approximate count of distinct elements.

Example

s = Series(rand(1:10, 1000), HyperLogLog(12))
value(s)

source

OnlineStats.KMeans — Type.

KMeans(p, k)

Approximate K-Means clustering of k clusters and p variables.

Example

using OnlineStats, Distributions
d = MixtureModel([Normal(0), Normal(5)])
y = rand(d, 100_000, 1)
s = Series(y, LearningRate(.6), KMeans(1, 2))

source

OnlineStats.Lag — Type.

Lag(b, T = Float64)

Store the last b values for a data stream of type T.

source

OnlineStats.LinReg — Type.

LinReg(p, λ::Float64 = 0.0)  # use λ for all parameters
LinReg(p, λfactor::Vector{Float64})

Ridge regression of p variables with elementwise regularization.

Example

x = randn(100, 10)
y = x * linspace(-1, 1, 10) + randn(100)
o = LinReg(10)
Series((x,y), o)
value(o)

source

OnlineStats.LinRegBuilder — Type.

LinRegBuilder(p)

Create an object from which any variable can be regressed on any other set of variables, optionally with ridge (PenaltyFunctions.L2Penalty) regularization. The main function to use with LinRegBuilder is coef:

coef(o::LinRegBuilder, λ = 0; y=1, x=[2,3,...], bias=true, verbose=false)

Return the coefficients of a regressing column y on columns x with ridge (L2Penalty) parameter λ. An intercept (bias) term is added by default.

Examples

x = randn(1000, 10)
o = LinRegBuilder(10)
s = Series(x, o)

# let response = x[:, 3]
coef(o; y=3, verbose=true) 

# let response = x[:, 7], predictors = x[:, [2, 5, 4]]
coef(o; y = 7, x = [2, 5, 4]) 

#

source

OnlineStats.MSPI — Type.

MSPI()  # Majorized stochastic proximal iteration
MSPI2()
OMAS()  # Online MM - Averaged Surrogate
OMAS2()
OMAP()  # Online MM - Averaged Parameter
OMAP2()

Updaters based on majorizing functions. MSPI/OMAS/OMAP define a family of algorithms and not a specific update, thus each type has two possible versions.

See https://arxiv.org/abs/1306.4650 for OMAS

Ask @joshday for details on OMAP and MSPI

source

OnlineStats.MV — Type.

MV(p, o)
p * o

Track p univariate OnlineStats o.

Example

y = randn(1000, 5)
o = MV(5, Mean())
s = Series(y, o)

Series(y, 5Mean())

source

OnlineStats.Mean — Type.

Mean()

Univariate mean.

Example

s = Series(randn(100), Mean())
value(s)

source

OnlineStats.Moments — Type.

Moments()

First four non-central moments.

Example

s = Series(randn(1000), Moments(10))
value(s)

source

OnlineStats.NADAM — Type.

NADAM(α1 = .99, α2 = .999)

Adaptive Moment Estimation with momentum parameters α1 and α2.

source

OnlineStats.NBClassifier — Type.

NBClassifier(p, T, b = 20)

Create a Naive Bayes classifier for p predictors for classes of type T. Conditional probabilities are estimated using the Hist (with AdaptiveBins) type with b bins.

Example

x = randn(100, 5)
y = rand(Bool, 100)
o = NBClassifier(5, Bool)
Series((x,y), o)
predict(o, x)
classify(o,x)

source

OnlineStats.NSGD — Type.

NSGD(α)

Nesterov accelerated Proximal Stochastic Gradient Descent.

source

OnlineStats.OMAP — Type.

MSPI()  # Majorized stochastic proximal iteration
MSPI2()
OMAS()  # Online MM - Averaged Surrogate
OMAS2()
OMAP()  # Online MM - Averaged Parameter
OMAP2()

Updaters based on majorizing functions. MSPI/OMAS/OMAP define a family of algorithms and not a specific update, thus each type has two possible versions.

See https://arxiv.org/abs/1306.4650 for OMAS

Ask @joshday for details on OMAP and MSPI

source

OnlineStats.OMAS — Type.

MSPI()  # Majorized stochastic proximal iteration
MSPI2()
OMAS()  # Online MM - Averaged Surrogate
OMAS2()
OMAP()  # Online MM - Averaged Parameter
OMAP2()

Updaters based on majorizing functions. MSPI/OMAS/OMAP define a family of algorithms and not a specific update, thus each type has two possible versions.

See https://arxiv.org/abs/1306.4650 for OMAS

Ask @joshday for details on OMAP and MSPI

source

OnlineStats.OrderStats — Type.

OrderStats(b)

Average order statistics with batches of size b. Ignores weight.

Example

s = Series(randn(1000), OrderStats(10))
value(s)

source

OnlineStats.PQuantile — Type.

PQuantile(τ = 0.5)

Calculate the approximate quantile via the P^2 algorithm. It is more computationally expensive than the algorithms used by Quantile, but also more exact.

Ref: https://www.cse.wustl.edu/~jain/papers/ftp/psqr.pdf

Example

y = randn(10^6)
o1, o2, o3 = PQuantile(.25), PQuantile(.5), PQuantile(.75)
s = Series(y, o1, o2, o3)
value(s)
quantile(y, [.25, .5, .75])

source

OnlineStats.Partition — Type.

Partition(o::OnlineStat, b::Int)

Incrementally partition a data stream where between b and 2b sections are summarized by o.

Example

using Plots
s = Series(cumsum(randn(10^6)), Partition(Mean()))
plot(s)

source

OnlineStats.Quantile — Type.

Quantile(q = [.25, .5, .75], alg = OMAS())

Approximate the quantiles q via the stochastic approximation algorithm alg. Options are SGD, MSPI, and OMAS. In practice, SGD and MSPI only work well when the variance of the data is small.

Example

y = randn(10_000)
τ = collect(.1:.1:.0)
Series(y, Quantile(τ, SGD()), Quantile(τ, MSPI()), Quantile(τ, OMAS()))

source

OnlineStats.RMSPROP — Type.

RMSPROP(α = .9)

source

OnlineStats.ReservoirSample — Type.

ReservoirSample(k, t = Float64)

Reservoir sample of k items.

Example

o = ReservoirSample(k, Int)
s = Series(o)
fit!(s, 1:10000)

source

OnlineStats.SGD — Type.

SGD()

Stochastic gradient descent.

source

OnlineStats.Series — Type.

Series(stats...)
Series(weight, stats...)
Series(data, weight, stats...)
Series(data, stats...)
Series(weight, data, stats...)

Track any number of OnlineStats.

Example

Series(Mean())
Series(randn(100), Mean())
Series(randn(100), ExponentialWeight(), Mean())

s = Series(Quantile([.25, .5, .75]))
fit!(s, randn(1000))

source

OnlineStats.StatLearn — Type.

StatLearn(p::Int, args...)

Fit a statistical learning model of p independent variables for a given loss, penalty, and λ. Additional arguments can be given in any order (and is still type stable):

loss = .5 * L2DistLoss(): any Loss from LossFunctions.jl

penalty = L2Penalty(): any Penalty (which has a prox method) from PenaltyFunctions.jl.

λ = fill(.1, p): a Vector of element-wise regularization parameters

updater = SGD(): SGD, ADAGRAD, ADAM, ADAMAX, MSPI

Details

The (offline) objective function that StatLearn approximately minimizes is

$\frac{1}{n}\sum_{i=1}^n f_i(\beta) + \sum_{j=1}^p \lambda_j g(\beta_j),$

where the $f_i$'s are loss functions evaluated on a single observation, $g$ is a penalty function, and the $\lambda_j$s are nonnegative regularization parameters.

Example

using LossFunctions, PenaltyFunctions
x = randn(100_000, 10)
y = x * linspace(-1, 1, 10) + randn(100_000)
o = StatLearn(10, .5 * L2DistLoss(), L1Penalty(), fill(.1, 10), SGD())
s = Series(o)
fit!(s, x, y)
coef(o)
predict(o, x)

source

OnlineStats.Sum — Type.

Sum()

Track the overall sum.

Example

s = Series(randn(1000), Sum())
value(s)

source

OnlineStats.Variance — Type.

Variance()

Univariate variance.

Example

s = Series(randn(100), Variance())
value(s)

source

OnlineStats.mapblocks — Function.

mapblocks(f::Function, b::Int, data, dim::ObsDimension = Rows())

Map data in batches of size b to the function f. If data includes an AbstractMatrix, the batches will be based on rows or columns, depending on dim. Most usage is through Julia's do block syntax.

Examples

s = Series(Mean())
mapblocks(10, randn(100)) do yi
    fit!(s, yi)
    info("nobs: $(nobs(s))")
end

x = [1 2 3 4; 
     1 2 3 4; 
     1 2 3 4;
     1 2 3 4]
mapblocks(println, 2, x)
mapblocks(println, 2, x, Cols())

source

StatsBase.confint — Function.

confint(b::Bootstrap, coverageprob = .95)

Return a confidence interval for a Bootstrap b.

source

StatsBase.fit! — Method.

fit!(s::Series, data)

Update a Series with more data.

Examples

# Univariate Series 
s = Series(Mean())
fit!(s, randn(100))

# Multivariate Series
x = randn(100, 3)
s = Series(CovMatrix(3))
fit!(s, x)  # Same as fit!(s, x, Rows())
fit!(s, x', Cols())

# Model Series
x, y = randn(100, 10), randn(100)
s = Series(LinReg(10))
fit!(s, (x, y))

source

OnlineStats.BiasVec — Type.

BiasVec(x, bias = 1.0)

Lightweight wrapper of a vector which adds a "bias" term at the end.

Example

OnlineStats.BiasVec(rand(5), 10)

source

OnlineStats.KnownBins — Type.

Calculate a histogram over a fixed range.

source

OnlineStats.MSPI2 — Type.

MSPI()  # Majorized stochastic proximal iteration
MSPI2()
OMAS()  # Online MM - Averaged Surrogate
OMAS2()
OMAP()  # Online MM - Averaged Parameter
OMAP2()

Updaters based on majorizing functions. MSPI/OMAS/OMAP define a family of algorithms and not a specific update, thus each type has two possible versions.

See https://arxiv.org/abs/1306.4650 for OMAS

Ask @joshday for details on OMAP and MSPI

source

OnlineStats.OMAP2 — Type.

MSPI()  # Majorized stochastic proximal iteration
MSPI2()
OMAS()  # Online MM - Averaged Surrogate
OMAS2()
OMAP()  # Online MM - Averaged Parameter
OMAP2()

Updaters based on majorizing functions. MSPI/OMAS/OMAP define a family of algorithms and not a specific update, thus each type has two possible versions.

See https://arxiv.org/abs/1306.4650 for OMAS

Ask @joshday for details on OMAP and MSPI

source

OnlineStats.OMAS2 — Type.

MSPI()  # Majorized stochastic proximal iteration
MSPI2()
OMAS()  # Online MM - Averaged Surrogate
OMAS2()
OMAP()  # Online MM - Averaged Parameter
OMAP2()

Updaters based on majorizing functions. MSPI/OMAS/OMAP define a family of algorithms and not a specific update, thus each type has two possible versions.

See https://arxiv.org/abs/1306.4650 for OMAS

Ask @joshday for details on OMAP and MSPI

source

Base.merge! — Method.

merge!(s1::Series, s2::Series, arg)

Merge s2 into s1 in place where s2's influence is determined by arg. Options for arg` are:

:append (default)

appends2tos1` with influence determined by number of observations. For

EqualWeight, this is equivalent to fit!(s1, data2) where s2 = Series(data2, o...).

:mean

Use the average (weighted by nobs) of the Series' generated weights.

:singleton

treat s2 as a single observation.

any Float64 in [0, 1]

source

Base.merge — Method.

See merge!

source

OnlineStats.always — Method.

always returns true

source

OnlineStatsBase.Bounded — Type.

Bounded(w::Weight, λ::Float64)

Bound the weight by a constant.

$γ_{bounded}(t) = max(γ(t), λ)$

Example

Bounded(EqualWeight(), .1)

source

OnlineStatsBase.EqualWeight — Type.

EqualWeight()

Equally weighted observations.

$γ(t) = 1 / t$

Example

Series(randn(100), EqualWeight(), Variance())

source

OnlineStatsBase.ExactStat — Type.

An OnlineStat which can be updated exactly. Subtypes of ExactStat use EqualWeight() as the default weight.

source

OnlineStatsBase.ExponentialWeight — Type.

ExponentialWeight(λ::Float64)
ExponentialWeight(lookback::Int)

Exponentially weighted observations.

$γ(t) = λ = 2 / (lookback + 1)$

Example

Series(randn(100), ExponentialWeight(), Variance())

source

OnlineStatsBase.HarmonicWeight — Type.

HarmonicWeight(a = 10.0)

Weight determined by harmonic series.

$γ(t) = a / (a + t - 1)$

Example

Series(randn(1000), HarmonicWeight(), QuantileMSPI())

source

OnlineStatsBase.LearningRate — Type.

LearningRate(r = .6)

Slowly decreasing weight.

$γ(t) = 1 / t^r$

Example

Series(randn(1000), LearningRate(.7), QuantileMM(), QuantileMSPI(), QuantileSGD())

source

OnlineStatsBase.LearningRate2 — Type.

LearningRate2(c = .5)

Slowly decreasing weight.

$γ(t) = 1 / (1 + c * (t - 1))$

Example

Series(randn(1000), LearningRate2(.3), QuantileMM(), QuantileMSPI(), QuantileSGD())

source

OnlineStatsBase.McclainWeight — Type.

McclainWeight(α = .1)

Weight which decreases into a constant.

$γ(t) = γ(t-1) / (1 + γ(t) - α)$

Example

Series(randn(100), McclainWeight(), Mean())

source

OnlineStatsBase.Scaled — Type.

Scaled(w::Weight, λ::Float64)

Scale a weight by a constant.

$γ_{scaled}(t) = λ * γ(t)$

Example

Bounded(LearningRate(.5), .1)

Series(randn(1000), 2.0 * LearningRate(.9), QuantileMM())

source

OnlineStatsBase.StochasticStat — Type.

An OnlineStat which must be approximated. Subtypes of StochasticStat use LearningRate() as the default weight. Additionally, subtypes should be parameterized by an algorithm, which is an optional last argument. For example:

struct Quantile{T <: Updater} <: StochasticStat{0}
    value::Vector{Float64}
    τ::Vector{Float64}
    updater::T 
end
Quantile(τ::AbstractVector = [.25, .5, .75], u::Updater = SGD()) = ...

source

OnlineStatsBase.Weight — Type.

Subtypes of Weight must be callable to produce the weight given the current number of observations in an OnlineStat n and the number of new observations (n2).

MyWeight(n, n2 = 1)

source

OnlineStatsBase._fit! — Function.

_fit!(o::OnlineStat, observation::TypeOfObservation, γ::Float64)

Update an OnlineStat with a single observation by a weight γ. TypeOfObservation depends on typof(o).

source