API

API

OnlineStats.ADADELTA

OnlineStats.ADAGRAD

OnlineStats.ADAM

OnlineStats.ADAMAX

OnlineStats.FitBeta

OnlineStats.FitCategorical

OnlineStats.FitCauchy

OnlineStats.FitGamma

OnlineStats.FitLogNormal

OnlineStats.FitMultinomial

OnlineStats.FitMvNormal

OnlineStats.FitNormal

OnlineStats.IHistogram

OnlineStats.LinRegBuilder

OnlineStats.MSPIQ

OnlineStats.NADAM

OnlineStats.NSGD

OnlineStats.OMASQ

OnlineStats.SGD

OnlineStats.StatLearn

OnlineStatsBase.Bootstrap

OnlineStatsBase.Bounded

OnlineStatsBase.CStat

OnlineStatsBase.CovMatrix

OnlineStatsBase.Diff

OnlineStatsBase.EqualWeight

OnlineStatsBase.ExponentialWeight

OnlineStatsBase.Extrema

OnlineStatsBase.HarmonicWeight

OnlineStatsBase.HyperLogLog

OnlineStatsBase.KMeans

OnlineStatsBase.LearningRate

OnlineStatsBase.LearningRate2

OnlineStatsBase.LinReg

OnlineStatsBase.MV

OnlineStatsBase.McclainWeight

OnlineStatsBase.Mean

OnlineStatsBase.Moments

OnlineStatsBase.OHistogram

OnlineStatsBase.OrderStats

OnlineStatsBase.QuantileMM

OnlineStatsBase.QuantileMSPI

OnlineStatsBase.QuantileSGD

OnlineStatsBase.ReservoirSample

OnlineStatsBase.Scaled

OnlineStatsBase.Series

OnlineStatsBase.Sum

OnlineStatsBase.Variance

OnlineStatsBase.mapblocks

OnlineStatsBase.replicates

StatsBase.confint

StatsBase.fit!

OnlineStats.ADADELTA — Type.

ADADELTA(ρ = .95)

ADADELTA ignores weight.

source

OnlineStats.ADAGRAD — Type.

ADAGRAD()

Adaptive (element-wise learning rate) stochastic proximal gradient descent.

source

OnlineStats.ADAM — Type.

ADAM(α1, α2)

Adaptive Moment Estimation with momentum parameters α1 and α2.

source

OnlineStats.ADAMAX — Type.

ADAMAX(η, β1, β2)

ADAMAX with step size η and momentum parameters β1, β2

source

OnlineStats.FitBeta — Type.

FitBeta()

Online parameter estimate of a Beta distribution (Method of Moments).

using Distributions, OnlineStats
y = rand(Beta(3, 5), 1000)
o = FitBeta()
s = Series(y, o)
Beta(value(o)...)

source

OnlineStats.FitCategorical — Type.

FitCategorical(T)

Fit a categorical distribution where the inputs are of type T.

using Distributions
s = Series(rand(1:10, 1000), FitCategorical(Int))
value(s)

vals = ["small", "medium", "large"]
o = FitCategorical(String)
s = Series(rand(vals, 1000), o)
value(o)

source

OnlineStats.FitCauchy — Type.

FitCauchy()

Online parameter estimate of a Cauchy distribution.

using Distributions
y = rand(Cauchy(0, 10), 10_000)
o = FitCauchy()
s = Series(y, o)
Cauchy(value(o)...)

source

OnlineStats.FitGamma — Type.

FitGamma()

Online parameter estimate of a Gamma distribution (Method of Moments).

using Distributions
y = rand(Gamma(5, 1), 1000)
o = FitGamma()
s = Series(y, o)
Gamma(value(o)...)

source

OnlineStats.FitLogNormal — Type.

FitLogNormal()

Online parameter estimate of a LogNormal distribution (MLE).

using Distributions
y = rand(LogNormal(3, 4), 1000)
o = FitLogNormal()
s = Series(y, o)
LogNormal(value(o)...)

source

OnlineStats.FitMultinomial — Type.

FitMultinomial(p)

Online parameter estimate of a Multinomial distribution. The sum of counts does not need to be consistent across observations. Therefore, the n parameter of the Multinomial distribution is returned as 1.

using Distributions
y = rand(Multinomial(10, [.2, .2, .6]), 1000)
o = FitMultinomial(3)
s = Series(y', o)
Multinomial(value(o)...)

source

OnlineStats.FitMvNormal — Type.

FitMvNormal(d)

Online parameter estimate of a d-dimensional MvNormal distribution (MLE).

using Distributions
y = rand(MvNormal(zeros(3), eye(3)), 1000)
o = FitMvNormal(3)
s = Series(y', o)
MvNormal(value(o)...)

source

OnlineStats.FitNormal — Type.

FitNormal()

Online parameter estimate of a Normal distribution (MLE).

using Distributions
y = rand(Normal(-3, 4), 1000)
o = FitNormal()
s = Series(y, o)
Normal(value(o)...)

source

OnlineStats.IHistogram — Type.

IHistogram(b::Integer)

Incrementally construct a histogram of b bins.

Example

using Plots
o = IHistogram(1000)
Series(randexp(100_000), o)
plot(o)

source

OnlineStats.LinRegBuilder — Type.

LinRegBuilder(p)

Create an object from which any variable can be regressed on any other set of variables.

Example

x = randn(1000, 10)
o = LinRegBuilder(10)
s = Series(x, o)

# let response = x[:, 3], predictors = x[:, setdiff(1:10, 3)]
coef(o, 3) 

# let response = x[:, 7], predictors = x[:, [2, 5, 4]]
coef(o, 7, [2, 5, 4])

source

OnlineStats.MSPIQ — Type.

MSPIQ()

MSPI-Q algorithm using a Lipschitz constant to majorize the objective.

source

OnlineStats.NADAM — Type.

NADAM(α1, α2)

Adaptive Moment Estimation with momentum parameters α1 and α2.

source

OnlineStats.NSGD — Type.

NSGD(α)

Nesterov accelerated Proximal Stochastic Gradient Descent.

source

OnlineStats.OMASQ — Type.

Experimental: OMM-constant

source

OnlineStats.SGD — Type.

SGD()

Proximal Stochastic Gradient Descent.

source

OnlineStats.StatLearn — Type.

StatLearn(p::Int, args...)

Fit a statistical learning model of p independent variables for a given loss, penalty, and λ. Additional arguments can be given in any order (and is still type stable):

loss = .5 * L2DistLoss(): any Loss from LossFunctions.jl

penalty = L2Penalty(): any Penalty (which has a prox method) from PenaltyFunctions.jl.

λ = fill(.1, p): a Vector of element-wise regularization parameters

updater = SGD(): SGD, ADAGRAD, ADAM, ADAMAX

Details

The (offline) objective function which StatLearn approximately minimizes is

$\frac{1}{n}\sum_{i=1}^n f_i(\beta) + \sum_{j=1}^p \lambda_j g(\beta_j),$

where the $f_i$'s are loss functions evaluated on a single observation, $g$ is a penalty function, and the $\lambda_j$s are nonnegative regularization parameters.

Example

using LossFunctions, PenaltyFunctions
x = randn(100_000, 10)
y = x * linspace(-1, 1, 10) + randn(100_000)
o = StatLearn(10, .5 * L2DistLoss(), L1Penalty(), fill(.1, 10), SGD())
s = Series(o)
fit!(s, x, y)
coef(o)
predict(o, x)

source

OnlineStatsBase.Bootstrap — Type.

Bootstrap(o::OnlineStat, nreps = 100, d = [0, 2], f = value)

Online Statistical Bootstrapping. Create nreps replicates of the OnlineStat o. When fit! is called, each of the replicates will be updated rand(d) times. value(b::Bootstrap) returns f mapped to the replicates.

Example

b = Bootstrap(Mean())
fit!(b, randn(1000))
value(b)        # `f` mapped to replicates
mean(value(b))  # mean

source

OnlineStatsBase.Bounded — Type.

Bounded(weight, λ)

Give a Weight a lower bound.

source

OnlineStatsBase.CStat — Type.

CStat(stat)

Track a univariate OnlineStat for complex numbers. A copy of stat is made to separately track the real and imaginary parts.

Example

y = randn(100) + randn(100)im
Series(y, CStat(Mean()))

source

OnlineStatsBase.CovMatrix — Type.

CovMatrix(d)

Covariance Matrix of d variables.

Example

y = randn(100, 5)
Series(y, CovMatrix(5))

source

OnlineStatsBase.Diff — Type.

Diff()

Track the difference and the last value.

Example

s = Series(randn(1000), Diff())
value(s)

source

OnlineStatsBase.EqualWeight — Type.

EqualWeight()

Equally weighted observations

Weight at observation t is γ = 1 / t

source

OnlineStatsBase.ExponentialWeight — Type.

ExponentialWeight(λ::Real = 0.1)
ExponentialWeight(lookback::Integer)

Exponentially weighted observations (constant)

Weight at observation t is γ = λ

source

OnlineStatsBase.Extrema — Type.

Extrema()

Maximum and minimum.

Example

s = Series(randn(100), Extrema())
value(s)

source

OnlineStatsBase.HarmonicWeight — Type.

HarmonicWeight(a = 10.0)

Decreases at a slow rate

Weight at observation t is γ = a / (a + t - 1)

source

OnlineStatsBase.HyperLogLog — Type.

HyperLogLog(b)  # 4 ≤ b ≤ 16

Approximate count of distinct elements.

Example

s = Series(rand(1:10, 1000), HyperLogLog(12))
value(s)

source

OnlineStatsBase.KMeans — Type.

KMeans(p, k)

Approximate K-Means clustering of k clusters and p variables.

Example

using OnlineStats, Distributions
d = MixtureModel([Normal(0), Normal(5)])
y = rand(d, 100_000, 1)
s = Series(y, LearningRate(.6), KMeans(1, 2))

source

OnlineStatsBase.LearningRate — Type.

LearningRate(r = .6)

Mainly for stochastic approximation types

Decreases at a "slow" rate

Weight at observation t is γ = 1 / t ^ r

source

OnlineStatsBase.LearningRate2 — Type.

LearningRate2(c = .5)

Mainly for stochastic approximation types

Decreases at a "slow" rate

Weight at observation t is γ = inv(1 + c * (t - 1))

source

OnlineStatsBase.LinReg — Type.

LinReg(p, λ::Float64 = 0.0)  # use λ for all parameters
LinReg(p, λfactor::Vector{Float64})

Ridge regression of p variables with elementwise regularization.

Example

x = randn(100, 10)
y = x * linspace(-1, 1, 10) + randn(100)
o = LinReg(10)
Series((x,y), o)
value(o)

source

OnlineStatsBase.MV — Type.

MV(p, o)

Track p univariate OnlineStats o.

Example

y = randn(1000, 5)
o = MV(5, Mean())
s = Series(y, o)

source

OnlineStatsBase.McclainWeight — Type.

McclainWeight(ᾱ = 0.1)

"smoothed" version of BoundedEqualWeight

weights asymptotically approach ᾱ

Weight at observation t is γ(t-1) / (1 + γ(t-1) - ᾱ)

source

OnlineStatsBase.Mean — Type.

Mean()

Univariate mean.

Example

s = Series(randn(100), Mean())
value(s)

source

OnlineStatsBase.Moments — Type.

Moments()

First four non-central moments.

Example

s = Series(randn(1000), Moments(10))
value(s)

source

OnlineStatsBase.OHistogram — Type.

OHistogram(range)

Make a histogram with bins given by range. Uses left-closed bins.

Example

y = randn(100)
s = Series(y, OHistogram(-4:.1:4))
value(s)

source

OnlineStatsBase.OrderStats — Type.

OrderStats(b)

Average order statistics with batches of size b. Ignores weight.

Example

s = Series(randn(1000), OrderStats(10))
value(s)

source

OnlineStatsBase.QuantileMM — Type.

QuantileMM(q = [.25, .5, .75])

Approximate quantiles via an online MM algorithm (OMAS).

Example

s = Series(randn(1000), QuantileMM())
value(s)

source

OnlineStatsBase.QuantileMSPI — Type.

QuantileMSPI(q = [.25, .5, .75])

Approximate quantiles via Majorized Stochastic Proximal Iteration (MSPI).

Example

s = Series(randn(1000), QuantileMSPI())
value(s)

source

OnlineStatsBase.QuantileSGD — Type.

QuantileSGD(q = [.25, .5, .75])

Approximate quantiles via an stochastic subgradient descent.

Example

s = Series(randn(1000), QuantileSGD())
value(s)

source

OnlineStatsBase.ReservoirSample — Type.

ReservoirSample(k, t = Float64)

Reservoir sample of k items.

Example

o = ReservoirSample(k, Int)
s = Series(o)
fit!(s, 1:10000)

source

OnlineStatsBase.Scaled — Type.

Scaled(weight, λ)
λ * weight

Scale a weight by a constant.

source

OnlineStatsBase.Series — Type.

Series(stats...)
Series(weight, stats...)
Series(data, weight, stats...; dim = Rows())
Series(weight, data, stats...; dim = Rows())

Track any number of OnlineStats using a given weighting mechanism.

Examples

s = Series(ExponentialWeight(.1), Mean(), Variance(), Moments())
fit!(s, randn(1000))
value(s)

o = CovMatrix(5)
s = Series(randn(5, 1000), o; dim = Cols())
cor(o)

x, y = randn(1000, 10), randn(1000)
s = Series(LinReg(10))
fit!(s, (x, y))  # or fit!(s, x, y)
value(s)

source

OnlineStatsBase.Sum — Type.

Sum()

Track the overall sum.

Example

s = Series(randn(1000), Sum())
value(s)

source

OnlineStatsBase.Variance — Type.

Variance()

Univariate variance.

Example

s = Series(randn(100), Variance())
value(s)

source

OnlineStatsBase.mapblocks — Function.

mapblocks(f::Function, b::Int, data, dim::ObsDimension = Rows())

Map data in batches of size b to the function f. If data includes an AbstractMatrix, the batches will be based on rows or columns, depending on dim. Most usage is through Julia's do block syntax.

Examples

s = Series(Mean())
mapblocks(10, randn(100)) do yi
    fit!(s, yi)
    info("nobs: $(nobs(s))")
end

x = [1 2 3 4; 
     1 2 3 4; 
     1 2 3 4;
     1 2 3 4]
mapblocks(println, 2, x)
mapblocks(println, 2, x, Cols())

source

OnlineStatsBase.replicates — Method.

replicates(b)

Return the vector of replicates from Bootstrap b

source

StatsBase.confint — Function.

confint(b, coverageprob = .95)

Return a confidence interval for a Bootstrap b.

source

StatsBase.fit! — Function.

fit!(s::Series, data)
fit!(s::Series, data, w::StatsBase.AbstractWeights)

Update a Series with more data, optionally overriding the Weight.

Example

y = randn(100)
w = rand(100)

s = Series(Mean())
fit!(s, y[1])          # one observation: use Series weight
fit!(s, y[1], w[1])     # one observation: override weight
fit!(s, y)              # multiple observations: use Series weight
fit!(s, y, w[1])        # multiple observations: override each weight with w[1]
fit!(s, y, Weights(w))  # multiple observations: y[i] uses weight w[i]

x, y = randn(100, 5), randn(100)
s = Series(LinReg(5))
fit!(s, (x, y))  # or fit!(s, x, y)

source