API

API

OnlineStats.ADADELTA

OnlineStats.ADAGRAD

OnlineStats.ADAM

OnlineStats.ADAMAX

OnlineStats.FitBeta

OnlineStats.FitCategorical

OnlineStats.FitCauchy

OnlineStats.FitGamma

OnlineStats.FitLogNormal

OnlineStats.FitMultinomial

OnlineStats.FitMvNormal

OnlineStats.FitNormal

OnlineStats.MSPIC

OnlineStats.MSPIF

OnlineStats.NADAM

OnlineStats.NSGD

OnlineStats.OMASQ

OnlineStats.OMASQF

OnlineStats.SGD

OnlineStats.StatLearn

OnlineStatsBase.Bootstrap

OnlineStatsBase.Bounded

OnlineStatsBase.CovMatrix

OnlineStatsBase.Diff

OnlineStatsBase.EqualWeight

OnlineStatsBase.ExponentialWeight

OnlineStatsBase.Extrema

OnlineStatsBase.HarmonicWeight

OnlineStatsBase.HyperLogLog

OnlineStatsBase.KMeans

OnlineStatsBase.LearningRate

OnlineStatsBase.LearningRate2

OnlineStatsBase.LinReg

OnlineStatsBase.MV

OnlineStatsBase.McclainWeight

OnlineStatsBase.Mean

OnlineStatsBase.Moments

OnlineStatsBase.OHistogram

OnlineStatsBase.OrderStats

OnlineStatsBase.QuantileMM

OnlineStatsBase.ReservoirSample

OnlineStatsBase.Scaled

OnlineStatsBase.Series

OnlineStatsBase.Sum

OnlineStatsBase.Variance

OnlineStats.maprows

OnlineStatsBase.replicates

StatsBase.confint

StatsBase.fit!

OnlineStats.ADADELTA — Type.

ADADELTA(ρ = .95)

ADADELTA ignores weight.

source

OnlineStats.ADAGRAD — Type.

ADAGRAD()

Adaptive (element-wise learning rate) stochastic proximal gradient descent.

source

OnlineStats.ADAM — Type.

ADAM(α1, α2)

Adaptive Moment Estimation with momentum parameters α1 and α2.

source

OnlineStats.ADAMAX — Type.

ADAMAX(η, β1, β2)

ADAMAX with step size η and momentum parameters β1, β2

source

OnlineStats.FitBeta — Type.

FitBeta()

Online parameter estimate of a Beta distribution (Method of Moments).

using Distributions, OnlineStats
y = rand(Beta(3, 5), 1000)
s = Series(y, FitBeta())
Beta(value(s)...)

source

OnlineStats.FitCategorical — Type.

FitCategorical(T)

Fit a categorical distribution where the inputs are of type T.

using Distributions
s = Series(rand(1:10, 1000), FitCategorical(Int))
value(s)

vals = ["small", "medium", "large"]
s = Series(rand(vals, 1000), FitCategorical(String))
value(s)

source

OnlineStats.FitCauchy — Type.

FitCauchy()

Online parameter estimate of a Cauchy distribution.

using Distributions
y = rand(Cauchy(0, 10), 10_000)
s = Series(y, FitCauchy())
Cauchy(value(s)...)

source

OnlineStats.FitGamma — Type.

FitGamma()

Online parameter estimate of a Gamma distribution (Method of Moments).

using Distributions
y = rand(Gamma(5, 1), 1000)
s = Series(y, FitGamma())
Gamma(value(s)...)

source

OnlineStats.FitLogNormal — Type.

FitLogNormal()

Online parameter estimate of a LogNormal distribution (MLE).

using Distributions
y = rand(LogNormal(3, 4), 1000)
s = Series(y, FitLogNormal())
LogNormal(value(s)...)

source

OnlineStats.FitMultinomial — Type.

FitMultinomial(p)

Online parameter estimate of a Multinomial distribution.

using Distributions
y = rand(Multinomial(10, [.2, .2, .6]), 1000)
s = Series(y', FitMultinomial())
Multinomial(value(s)...)

source

OnlineStats.FitMvNormal — Type.

FitMvNormal(d)

Online parameter estimate of a d-dimensional MvNormal distribution (MLE).

using Distributions
y = rand(MvNormal(zeros(3), eye(3)), 1000)
s = Series(y', FitMvNormal(3))

source

OnlineStats.FitNormal — Type.

FitNormal()

Online parameter estimate of a Normal distribution (MLE).

using Distributions
y = rand(Normal(-3, 4), 1000)
s = Series(y, FitNormal())

source

OnlineStats.MSPIC — Type.

MSPIC()

MSPI-Q algorithm using a Lipschitz constant to majorize the objective.

source

OnlineStats.MSPIF — Type.

Experimental: MSPI-full matrix

source

OnlineStats.NADAM — Type.

NADAM(α1, α2)

Adaptive Moment Estimation with momentum parameters α1 and α2.

source

OnlineStats.NSGD — Type.

NSGD(α)

Nesterov accelerated Proximal Stochastic Gradient Descent.

source

OnlineStats.OMASQ — Type.

Experimental: OMM-constant

source

OnlineStats.OMASQF — Type.

Experimental: OMM-full matrix

source

OnlineStats.SGD — Type.

SGD()

Proximal Stochastic Gradient Descent.

source

OnlineStats.StatLearn — Type.

StatLearn(p, loss, penalty, λ, updater)

Fit a statistical learning model of p independent variables for a given loss, penalty, and λ. Arguments are:

loss: any Loss from LossFunctions.jl

penalty: any Penalty from PenaltyFunctions.jl.

λ: a Vector of element-wise regularization parameters

updater: SPGD(), ADAGRAD(), ADAM(), or ADAMAX()

using LossFunctions, PenaltyFunctions x = randn(100_000, 10) y = x * linspace(-1, 1, 10) + randn(100_000) o = StatLearn(10, L2DistLoss(), L1Penalty(), fill(.1, 10), SPGD()) s = Series(o) fit!(s, x, y) coef(o) predict(o, x)

source

OnlineStats.maprows — Method.

maprows(f::Function, b::Integer, data...)

Map rows of data in batches of size b. Most usage is done through do blocks.

s = Series(Mean())
maprows(10, randn(100)) do yi
    fit!(s, yi)
    info("nobs: $(nobs(s))")
end

source

OnlineStatsBase.Bootstrap — Type.

Bootstrap(o::OnlineStat, nreps = 100, d = [0, 2], f = value)

Online Statistical Bootstrapping. Create nreps replicates of the OnlineStat o. When fit! is called, each of the replicates will be updated rand(d) times. value(b::Bootstrap) returns f mapped to the replicates.

b = Bootstrap(Mean())
fit!(b, randn(1000))
value(b)        # `f` mapped to replicates
mean(value(b))  # mean

source

OnlineStatsBase.Bounded — Type.

Bounded(weight, λ)

Give a Weight a lower bound.

source

OnlineStatsBase.CovMatrix — Type.

CovMatrix(d)

Covariance Matrix of d variables.

y = randn(100, 5)
Series(y, CovMatrix(5))

source

OnlineStatsBase.Diff — Type.

Diff()

Track the difference and the last value.

s = Series(randn(1000), Diff())
value(s)

source

OnlineStatsBase.EqualWeight — Type.

EqualWeight()

Equally weighted observations

Weight at observation t is γ = 1 / t

source

OnlineStatsBase.ExponentialWeight — Type.

ExponentialWeight(λ::Real = 0.1)
ExponentialWeight(lookback::Integer)

Exponentially weighted observations (constant)

Weight at observation t is γ = λ

source

OnlineStatsBase.Extrema — Type.

Extrema()

Maximum and minimum.

s = Series(randn(100), Extrema())
value(s)

source

OnlineStatsBase.HarmonicWeight — Type.

HarmonicWeight(a = 10.0)

Decreases at a slow rate

Weight at observation t is γ = a / (a + t - 1)

source

OnlineStatsBase.HyperLogLog — Type.

HyperLogLog(b)  # 4 ≤ b ≤ 16

Approximate count of distinct elements.

s = Series(rand(1:10, 1000), HyperLogLog(12))
value(s)

source

OnlineStatsBase.KMeans — Type.

KMeans(p, k)

Approximate K-Means clustering of k clusters and p variables.

using OnlineStats, Distributions
d = MixtureModel([Normal(0), Normal(5)])
y = rand(d, 100_000, 1)
s = Series(y, LearningRate(.6), KMeans(1, 2))

source

OnlineStatsBase.LearningRate — Type.

LearningRate(r = .6)

Mainly for stochastic approximation types

Decreases at a "slow" rate

Weight at observation t is γ = 1 / t ^ r

source

OnlineStatsBase.LearningRate2 — Type.

LearningRate2(c = .5)

Mainly for stochastic approximation types

Decreases at a "slow" rate

Weight at observation t is γ = inv(1 + c * (t - 1))

source

OnlineStatsBase.LinReg — Type.

LinReg(p, λ::Float64 = 0.0)  # use λ for all parameters
LinReg(p, λfactor::Vector{Float64})

Ridge regression of p variables with elementwise regularization.

x = randn(100, 10)
y = x * linspace(-1, 1, 10) + randn(100)
o = LinReg(10)
Series((x,y), o)
value(o)

source

OnlineStatsBase.MV — Type.

MV(p, o)

Track p univariate OnlineStats o y = randn(1000, 5) o = MV(5, Mean()) s = Series(y, o)

source

OnlineStatsBase.McclainWeight — Type.

McclainWeight(ᾱ = 0.1)

"smoothed" version of BoundedEqualWeight

weights asymptotically approach ᾱ

Weight at observation t is γ(t-1) / (1 + γ(t-1) - ᾱ)

source

OnlineStatsBase.Mean — Type.

Mean()

Univariate mean.

s = Series(randn(100), Mean())
value(s)

source

OnlineStatsBase.Moments — Type.

Moments()

First four non-central moments.

s = Series(randn(1000), Moments(10))
value(s)

source

OnlineStatsBase.OHistogram — Type.

OHistogram(range)

Make a histogram with bins given by range. Uses left-closed bins.

y = randn(100)
s = Series(y, OHistogram(-4:.1:4))
value(s)

source

OnlineStatsBase.OrderStats — Type.

OrderStats(b)

Average order statistics with batches of size b.

s = Series(randn(1000), OrderStats(10))
value(s)

source

OnlineStatsBase.QuantileMM — Type.

QuantileMM(q = 0.5)

Approximate quantiles via an online MM algorithm.

s = Series(randn(1000), QuantileMM())
value(s)

source

OnlineStatsBase.ReservoirSample — Type.

ReservoirSample(k, t = Float64)

Reservoir sample of k items.

o = ReservoirSample(k, Int)
s = Series(o)
fit!(s, 1:10000)

source

OnlineStatsBase.Scaled — Type.

Scaled(weight, λ)
λ * weight

Scale a weight by a constant.

source

OnlineStatsBase.Series — Type.

Series(stats...)
Series(weight, stats...)
Series(data, weight, stats...)
Series(weight, data, stats...)

Track any number of OnlineStats using a given weighting mechanism.

s = Series(ExponentialWeight(.1), Mean(), Variance(), Moments())
fit!(s, randn(1000))
value(s)

x, y = randn(1000, 10), randn(1000)
s = Series(LinReg(10))
fit!(s, (x,y))
value(s)

source

OnlineStatsBase.Sum — Type.

Sum()

Track the overall sum.

s = Series(randn(1000), Sum())
value(s)

source

OnlineStatsBase.Variance — Type.

Variance()

Univariate variance.

s = Series(randn(100), Variance())
value(s)

source

OnlineStatsBase.replicates — Method.

replicates(b)

Return the vector of replicates from Bootstrap b

source

StatsBase.fit! — Function.

fit!(s, y)
fit!(s, y, w)

Update a Series s with more data y and optional weighting w. y = randn(100) w = rand(100)

s = Series(Mean())
fit!(s, y[1])        # one observation: use Series weight
fit!(s, y[1], w[1])  # one observation: override weight
fit!(s, y)           # multiple observations: use Series weight
fit!(s, y, w[1])     # multiple observations: override each weight with w[1]
fit!(s, y, w)        # multiple observations: y[i] uses weight w[i]

source

StatsBase.confint — Function.

confint(b, coverageprob = .95)

Return a confidence interval for a Bootstrap b.

source